# v-cloud-test-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vibbraneo",
        "contact": {
            "email": "diogorcazevedo@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://administracaodosistema.com.br/sv/api"
        },
        {
            "url": "http://127.0.0.1:8000/api/"
        }
    ],
    "paths": {
...
        "/message/getAllFromDeal/{id}": {
            "get": {
                "tags": [
                    "message"
                ],
                "summary": "get all messages param Deal",
                "description": "List all messages from Deal",
                "operationId": "getAllMessagesFromDeal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/message/getAllFromDealAndUser/{deal_id}/{user_id}": {
            "get": {
                "tags": [
                    "message"
                ],
                "summary": "get all messages param Deal and user",
                "description": "List all messages from Deal and user",
                "operationId": "getAllMessagesFromDealAndUser",
                "parameters": [
                    {
                        "name": "deal",
                        "in": "path",
                        "description": "deal param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/message/getById/{id}": {
            "get": {
                "tags": [
                    "message"
                ],
                "summary": "get all message by id",
                "description": "List all message by id",
                "operationId": "getMessageById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/message/store/{deal_id}/{user_id}": {
            "post": {
                "tags": [
                    "message"
                ],
                "summary": "Mensagem do cliente",
                "description": "Mensagem do cliente interessado",
                "operationId": "messageStore",
                "parameters": [
                    {
                        "name": "deal_id",
                        "in": "path",
                        "description": "deal_id request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "id do usuário requerente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "store bid fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "value"
                                ],
                                "properties": {
                                    "deal_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2"
                                    },
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Proponho troca por iphone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/message/update/{id}": {
            "put": {
                "tags": [
                    "message"
                ],
                "summary": "Resposta do vendedor",
                "description": "Resposta do dono do negócio (vendedor)",
                "operationId": "messageUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do param",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass Update fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_id",
                                    "message"
                                ],
                                "properties": {
                                    "message_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Obrigado pelo interesse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}